+++ src/field/pokemon.ts
@@ -9,7 +9,7 @@ import type { AttackMoveResult } from "#app/@types/AttackMoveResult";
 import type { QueuedMove } from "#app/@types/QueuedMove";
 import type { StarterMoveset } from "#app/@types/StarterData";
 import type { TurnMove } from "#app/@types/TurnMove";
-import type { AnySound } from "#app/audioManager";
+import type { AnySound } from "#app/audio-manager";
 import { DYNAMAX_DAMAGE_TAKEN_FACTOR, PLAYER_PARTY_MAX_SIZE } from "#app/constants";
 import type { AbAttr } from "#app/data/ab-attrs/ab-attr";
 import type { Ability } from "#app/data/ability";
