+++ src/phases/evolution-phase.ts
@@ -3,7 +3,7 @@
 import { type FormChangePhase } from "#app/phases/form-change-phase";
 // -- end tsdoc imports --
 
-import type { AnySound } from "#app/audioManager";
+import type { AnySound } from "#app/audio-manager";
 import type { SpeciesFormEvolution } from "#app/data/balance/pokemon-evolutions";
 import { EVOLVE_MOVE } from "#app/data/balance/pokemon-level-moves";
 import type { PlayerPokemon, Pokemon } from "#app/field/pokemon";
